no curry operator
Decided not to have a built-in curry operator. It was a bit confused as it had to make assumptions as to what was an input to a function. In gall, all functions only have one input. Stuff like sum and pow take lists as that one input, but that is not universal.