function reduce
if a function is passed into a reduce, then the function is used to generate the values passed into the reducing function. the generating function takes only the index, and must either return a value for the given index, or undefined. When undefined is returned it then returns the accumulator.